refactor: one worker pool module, supervisor config object
#462 step 4 (structure consolidation): merge OnDemandWorkerManager +
WorkerSlotCoordinator into one worker-pool module — the slot state
machine (affinity -> create -> reclaim -> waiters) is one concept, and
the config-plumbing object between the two classes is gone. The pool
class is no longer exported; createWorkerManager(config, ports) is the
only construction path (namespace over classes). WorkerSupervisor now
takes a WorkerSupervisorOptions object instead of 7 positional args
(events stays required — no silent sink), and the three worker RPCs
(deliver/cancel/send) fold in from supervisor-client.ts, which is
deleted. Internal resolveDispatchTimeoutMs renamed to
resolveDeliverTimeoutMs (gate vocabulary). Behavior-preserving: all
merged method bodies are verbatim moves.
