+++ docs/implementation-status.md
@@ -102,11 +102,11 @@ The 2026-07-09 **determinism/verification round** (red-team + durable-execution 
 
 | Component | Status | Code | Notes |
 | --- | --- | --- | --- |
-| On-demand worker manager (spawn on demand, idle shutdown, max active) | ✅ | `packages/coordinator/src/worker-manager/manager.ts` | Used by `apps/server/src/execution/coordinator.ts`. #462 step 1: environment coupling is injected via a `WorkerPorts` object (`events: BusEvent.Sink` ledger edge — type lives in protocol — plus `toolRelay`/`inboundWait`); the coordinator package is session-free (protocol-only ring-2 driver, CI-ratcheted in `script/check-deps.ts`), and `apps/server/src/execution/coordinator.ts` is the composition-root binding of `events` to `Bus.publish` — the one binding P2 swaps when `Ledger.append` splits from lossy `Bus.publish`. #462 step 2: the driver verb is **`deliver(runId, task)`** (`cancel`/`send`/`stats`; no gate vocabulary in the public API), rejections are typed `WorkerDeliveryError` codes (`queue_full`/`shutting_down`/`duplicate_run`/`slot_wait_timeout`/`worker_restarted`) instead of string-matched `Error`s, and the command face is named `Execution.Driver` in protocol (`deliver` required; `send`/`cancel` capability-declared by presence) |
+| On-demand worker manager (spawn on demand, idle shutdown, max active) | ✅ | `packages/coordinator/src/worker-manager/worker-pool.ts` | Used by `apps/server/src/execution/coordinator.ts`. #462 step 1: environment coupling is injected via a `WorkerPorts` object (`events: BusEvent.Sink` ledger edge — type lives in protocol — plus `toolRelay`/`inboundWait`); the coordinator package is session-free (protocol-only ring-2 driver, CI-ratcheted in `script/check-deps.ts`), and `apps/server/src/execution/coordinator.ts` is the composition-root binding of `events` to `Bus.publish` — the one binding P2 swaps when `Ledger.append` splits from lossy `Bus.publish`. #462 step 2: the driver verb is **`deliver(runId, task)`** (`cancel`/`send`/`stats`; no gate vocabulary in the public API), rejections are typed `WorkerDeliveryError` codes (`queue_full`/`shutting_down`/`duplicate_run`/`slot_wait_timeout`/`worker_restarted`) instead of string-matched `Error`s, and the command face is named `Execution.Driver` in protocol (`deliver` required; `send`/`cancel` capability-declared by presence). #462 step 4: the slot state machine and delivery orchestration are one pool module (`worker-pool.ts`; the former `OnDemandWorkerManager`/`WorkerSlotCoordinator` split and its config-plumbing object are gone), the class is no longer exported — `createWorkerManager(config, ports)` is the only construction path |
 | Gate-side task policy stamping (`worker.spawn` → `PolicyPlan`) | ✅ | `packages/openomni/src/dispatch/handlers/worker.ts`, `packages/openomni/src/policy/resolver.ts`, `packages/openomni/src/dispatch/setup.ts` | #462 step 3 (§7): the gate resolves actor/target labels through `PolicyResolver` and stamps the resulting `policyPlan` (always including required `builtin:tool-permission` + `builtin:idle-nudge`) onto the spawn request; the plan rides the existing pipeline (`Execution.Request.policyPlan` → IPC `coordinator.spawn_run` → `worker-runner.ts` → `buildWorkerMiddleware` → worker `PolicyEngine`) as opaque data below the gate — no second authorization surface. Custom rules inject via `BuiltInDispatchOptions.policyResolver` at the composition root. Deferred: `Permission` stamping (the resolver produces policy IDs only; the fail-closed `builtin:tool-permission` policy is the enforcement lever) and connector-endpoint stamping (joins #462 driver-registry unification). The dispatch→driver session invariant is runtime-enforced: `sessionTreeId !== request.sessionId` now rejects with `WorkerDeliveryError code session_mismatch` |
 | Worker-local `child_agent` tool (same-runtime nested child agents) | ✅ | `apps/server/src/execution/worker-runner.ts`, `packages/openomni/src/execution-runtime/child-agent/`, `packages/openomni/src/execution-runtime/tool/agent/tools/child-agent.ts` | Worker-only tool registered by the server worker runner when requested. Child agents run in-process under the parent worker run, inherit worker context/middleware, can select only tools already exposed to the parent worker, cannot delegate further (`depth >= 1` removes delegation tools), emit lightweight lifecycle Bus events, can queue completion back to the parent run via `notifyOnComplete`, and are cancelled when the parent worker run finishes. This is a lightweight nested execution primitive, not a durable WorkerRun/SubagentRuntime replacement |
-| Worker supervisor internals | ✅ | `packages/coordinator/src/worker-supervision/supervisor.ts` | Shared process supervisor used by `worker-manager`; legacy `createWorkerPool` facade removed |
-| Session-to-slot affinity | ✅ | `packages/coordinator/src/worker-manager/worker-slot-coordinator.ts` | Plain per-coordinator affinity map (session → slot); the standalone SessionRouter module (least-load routing/refCounts) was deleted as dead API |
+| Worker supervisor internals | ✅ | `packages/coordinator/src/worker-supervision/supervisor.ts` | Shared process supervisor used by `worker-manager`; legacy `createWorkerPool` facade removed. #462 step 4: constructed from a `WorkerSupervisorOptions` object (the 7-positional constructor is gone; `events` stays required per the no-silent-sink rule) and the three worker RPCs (`deliver`/`cancel`/`send`) live on the supervisor directly — `supervisor-client.ts` is folded in |
+| Session-to-slot affinity | ✅ | `packages/coordinator/src/worker-manager/worker-pool.ts` | Plain per-pool affinity map (session → slot) inside the merged pool module; the standalone SessionRouter module (least-load routing/refCounts) was deleted as dead API |
 | Crash recovery (mark interrupted at boot) | ✅ | `apps/server/src/execution/recovery.ts` | Moved out of the coordinator in #462 step 1 — boot recovery is ledger work over session stores, not process-driver work, so it lives with the server composition root |
 | Injection queue (async delivery at turn.finish) | ✅ | `packages/openomni/src/execution-runtime/injection-queue.ts` | |
 | Cron job registry persistence | ✅ | `packages/openomni/src/execution-runtime/cron-job-registry.ts`, `packages/session/src/storage/sqlite-cron-job-adapter.ts` | `schedule.create` jobs persist in SQLite and survive `Storage` reinitialize; `schedule.cancel` removes persisted jobs |
